Fix: 로또 구입 금액 에러 수정 정수가 넘어가는 경우 예외 발생 제거

+++ src/main/java/domain/Money.java
@@ -3,11 +3,11 @@ package domain;
 import validator.MoneyValidator;
 
 public class Money {
-    private final int money;
+    private final long money;
 
     public Money(String money) {
         validate(money);
-        this.money = Integer.parseInt(money);
+        this.money = Long.parseLong(money);
     }
 
     private void validate(String money) {
@@ -16,7 +16,7 @@ public class Money {
         MoneyValidator.isThousandUnit(money);
     }
 
-    public int getMoney() {
+    public long getMoney() {
         return money;
     }
 }
